Move get_config_type_shape() to schema.utils
`get_config_type_shape()` is a pure schema object to EdgeQL AST
transformation, and since it's used outside of the compiler,
`schema.utils` seems like a better place for it.
