+++ edb/schema/utils.py
@@ -33,6 +33,7 @@ from . import name as sn
 from . import objects as so
 
 if TYPE_CHECKING:
+    from . import objtypes as s_objtypes
     from . import schema as s_schema
     from . import types as s_types
 
@@ -872,3 +873,75 @@ def const_ast_from_python(val: Any) -> qlast.BaseConstant:
         return qlast.BytesConstant.from_python(val)
     else:
         raise ValueError(f'unexpected constant type: {type(val)!r}')
+
+
+def get_config_type_shape(
+    schema: s_schema.Schema,
+    stype: s_objtypes.ObjectType,
+    path: List[qlast.Base],
+) -> List[qlast.ShapeElement]:
+    shape = []
+    seen: Set[str] = set()
+
+    stypes = [stype] + list(stype.descendants(schema))
+
+    for t in stypes:
+        t_name = t.get_name(schema)
+
+        for pn, p in t.get_pointers(schema).items(schema):
+            if pn in ('id', '__type__') or pn in seen:
+                continue
+
+            elem_path: List[qlast.Base] = []
+
+            if t is not stype:
+                elem_path.append(
+                    qlast.TypeIntersection(
+                        type=qlast.TypeName(
+                            maintype=qlast.ObjectRef(
+                                module=t_name.module,
+                                name=t_name.name,
+                            ),
+                        ),
+                    ),
+                )
+
+            elem_path.append(qlast.Ptr(ptr=qlast.ObjectRef(name=pn)))
+
+            ptype = p.get_target(schema)
+
+            if ptype.is_object_type():
+                subshape = get_config_type_shape(
+                    schema, ptype, path + elem_path)
+                subshape.append(
+                    qlast.ShapeElement(
+                        expr=qlast.Path(
+                            steps=[
+                                qlast.Ptr(
+                                    ptr=qlast.ObjectRef(name='_tname'),
+                                ),
+                            ],
+                        ),
+                        compexpr=qlast.Path(
+                            steps=path + elem_path + [
+                                qlast.Ptr(
+                                    ptr=qlast.ObjectRef(name='__type__')),
+                                qlast.Ptr(
+                                    ptr=qlast.ObjectRef(name='name')),
+                            ],
+                        ),
+                    ),
+                )
+            else:
+                subshape = []
+
+            shape.append(
+                qlast.ShapeElement(
+                    expr=qlast.Path(steps=elem_path),
+                    elements=subshape,
+                ),
+            )
+
+            seen.add(pn)
+
+    return shape
