+++ edb/server/config/spec.py
@@ -29,6 +29,7 @@ from edb.edgeql import codegen as qlcodegen
 from edb.edgeql import compiler as qlcompiler
 from edb.edgeql import qltypes
 from edb.ir import staeval
+from edb.schema import utils as s_utils
 
 from . import types
 
@@ -179,7 +180,7 @@ def generate_config_query(schema) -> str:
             expr=qlast.Path(
                 steps=[ref]
             ),
-            elements=qlcompiler.get_config_type_shape(schema, cfg, path=[ref]),
+            elements=s_utils.get_config_type_shape(schema, cfg, path=[ref]),
         ),
         limit=qlast.IntegerConstant(
             value='1',
